Complete main func
In the main function first we set all the memory
used by hashTable to '0', then using insert_person
function we add 2 persons to the hashTable.
finally we ask the user for a name and using the
search_person function we look up and in case of
person exsisting in the hashTable we print their
name and info. otherwise 'Person not found' will
be printed.